+++ src/routes/rootNavigation.tsx
@@ -3,9 +3,16 @@ import React from "react"
 
 export const navigationRef = React.createRef<NavigationContainerRef<any>>()
 
-export function navigate(name: string, params: object) {
+export function navigate(name: string, params?: object) {
   if (!navigationRef.current)
     return;
 
   navigationRef.current.navigate(name, params)
+}
+
+export function getRoutes() {
+  if (!navigationRef.current)
+    return;
+
+  return navigationRef.current.getState()?.routes
 }